Extracted Manage Sessions dialog into separate class
